# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sentinel Authorized Pentest Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scopes:
    get:
      operationId: listScopes
      tags: [scopes]
      summary: List authorization scopes
      responses:
        "200":
          description: List of scopes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scope"
    post:
      operationId: createScope
      tags: [scopes]
      summary: Create a new authorization scope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScopeInput"
      responses:
        "201":
          description: Created scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scope"
  /scopes/{id}:
    get:
      operationId: getScope
      tags: [scopes]
      summary: Get a scope by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scope"
        "404":
          description: Not found
    patch:
      operationId: updateScope
      tags: [scopes]
      summary: Update a scope
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScopeUpdate"
      responses:
        "200":
          description: Updated scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scope"
    delete:
      operationId: deleteScope
      tags: [scopes]
      summary: Delete a scope
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Targets ─────────────────────────────────────────────────────────────
  /targets:
    get:
      operationId: listTargets
      tags: [targets]
      summary: List targets
      parameters:
        - name: scopeId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of targets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Target"
    post:
      operationId: createTarget
      tags: [targets]
      summary: Add a target to a scope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "201":
          description: Created target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Target"
  /targets/{id}:
    get:
      operationId: getTarget
      tags: [targets]
      summary: Get a target by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Target"
        "404":
          description: Not found
    patch:
      operationId: updateTarget
      tags: [targets]
      summary: Update a target
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetUpdate"
      responses:
        "200":
          description: Updated target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Target"
    delete:
      operationId: deleteTarget
      tags: [targets]
      summary: Delete a target
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Scans ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Scope:
      type: object
      required: [id, name, signedBy, roeActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        signedBy:
          type: string
        roeActive:
          type: boolean
        validFrom:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        createdAt:
          type: string
        targetCount:
          type: integer
    ScopeInput:
      type: object
      required: [name, signedBy]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        signedBy:
          type: string
          minLength: 1
        roeActive:
          type: boolean
        validFrom:
          type: string
        validUntil:
          type: string
    ScopeUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        signedBy:
          type: string
        roeActive:
          type: boolean
        validFrom:
          type: string
        validUntil:
          type: string

    # ── Target ───────────────────────────────────────────────────────────
    Target:
      type: object
      required: [id, scopeId, host, allowed, activeModeEnabled, createdAt]
      properties:
        id:
          type: integer
        scopeId:
          type: integer
        scopeName:
          type: string
        host:
          type: string
        ip:
          type: ["string", "null"]
        portRange:
          type: ["string", "null"]
        allowed:
          type: boolean
        activeModeEnabled:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    TargetInput:
      type: object
      required: [scopeId, host]
      properties:
        scopeId:
          type: integer
        host:
          type: string
          minLength: 1
        ip:
          type: string
        portRange:
          type: string
        allowed:
          type: boolean
        activeModeEnabled:
          type: boolean
        notes:
          type: string
    TargetUpdate:
      type: object
      properties:
        host:
          type: string
        ip:
          type: string
        portRange:
          type: string
        allowed:
          type: boolean
        activeModeEnabled:
          type: boolean
        notes:
          type: string

    # ── Scan ─────────────────────────────────────────────────────────────
...