# lib/api-zod/src/generated/types/scopeUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel Authorized Pentest Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ScopeUpdate {
  name?: string;
  description?: string;
  signedBy?: string;
  roeActive?: boolean;
  validFrom?: string;
  validUntil?: string;
}