# lib/api-zod/src/generated/types/targetInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel Authorized Pentest Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TargetInput {
  scopeId: number;
  /** @minLength 1 */
  host: string;
  ip?: string;
  portRange?: string;
  allowed?: boolean;
  activeModeEnabled?: boolean;
  notes?: string;
}